test(stella-fleet): deduplicate the ScriptedGit test fake
git.rs's own test module and gc/tests.rs each declared an identical
GitHandler/ScriptedGit fake (only the doc comments differed), so
extending GitCli required editing both in sync. Move it into
git/test_support.rs, #[cfg(test)] pub(crate), and have both call sites
import it. manager() stays in git.rs's tests since it is specific to
those tests. Pure test-support refactor, no runtime behaviour change —
no witness test per CONTRIBUTING.md's rule for pure refactors.

Closes #3717
